requirements.txt
matplotlib
pandas
scipy-stack
scikit-image
pybalu
seaborn
librosa
pylint
autopep8